// src/clients/amazon.ts
class AmazonApi {
    api_key: string
    affiliate_id: string
    include_html: string = "false"
    output: "json" | "html" | "csv" = "json"
    exclude_sponsored: string = "true"
    api_url: string = 'https://api.asindataapi.com/request'
    constructor(api_key: string, affiliate_id: string) {
        this.api_key = api_key
        this.affiliate_id = affiliate_id
    }

    async search(amazon_domain: string, search_term: string): Promise<SearchResponse|null> {

        const callType = "search"

        // const response = await fetch(this.api_url, {
        //     method: "POST",
        //     headers: {
        //         "Content-Type": "application/json",
        //     },
        //     body: JSON.stringify({
        //         type: callType,
        //         api_key: this.api_key,
        //         associate_id: this.affiliate_id,
        //         amazon_domain: amazon_domain,
        //         search_term: search_term,
        //         exclude_sponsored: this.exclude_sponsored,
        //         include_html: this.include_html,
        //         output: this.output
        //     }),
        // });


        const queryParams = {
            type: callType,
            api_key: this.api_key,
            associate_id: this.affiliate_id,
            amazon_domain: amazon_domain,
            search_term: search_term,
            exclude_sponsored: this.exclude_sponsored,
            include_html: this.include_html,
            output: this.output
        };
        const queryString = new URLSearchParams(queryParams).toString();
        const urlWithQuery = `${this.api_url}?${queryString}`;
        const response = await fetch(urlWithQuery, {
            method: 'GET',
            headers: {
                "Content-Type": "application/json",
            }
        });

        let data = await response.json();

        if (response.ok) {
            return data as SearchResponse 
        } else {
            console.log("Request to api.asindataapi.com Failed.")
            return null
        }

    }

    countryToDomain(countryCode: string|null): string {
        if(countryCode && countryCode in countryDomainList){
            return countryDomainList[countryCode]
        }
        return countryDomainList.US
    }
}


export const amazon = new AmazonApi(
    process.env.ASINDATAAPI_KEY,
    process.env.AMAZON_AFFILIATE_ID
)